Refactor arizona-specific export code
- upgraded ai.lum.common dependency
